cleanup gnosis integration + fetch safe details from contract
