+++ package.json
@@ -9,6 +9,7 @@
     "@gnosis.pm/safe-apps-sdk": "^4.2.0",
     "@gnosis.pm/safe-apps-web3modal": "^2.0.0",
     "@gnosis.pm/safe-contracts": "^1.3.0",
+    "@gnosis.pm/safe-core-sdk": "^1.3.0",
     "@gnosis.pm/zodiac": "^1.0.3",
     "@walletconnect/web3-provider": "^1.6.5",
     "abi-decoder": "^2.4.0",
