+++ src/pages/MinionVault.jsx
@@ -10,7 +10,6 @@ import {
   Icon,
   Link as ChakraLink,
 } from '@chakra-ui/react';
-import { utils as Web3Utils } from 'web3';
 
 import { useToken } from '../contexts/TokenContext';
 import BalanceList from '../components/balanceList';
@@ -21,12 +20,10 @@ import MainViewLayout from '../components/mainViewLayout';
 import NftCard from '../components/nftCard';
 import TextBox from '../components/TextBox';
 import SafeMinionDetails from '../components/safeMinionDetails';
-import { chainByID } from '../utils/chain';
-import { isAmbModule } from '../utils/contract';
+import { fetchSafeDetails } from '../utils/gnosis';
 import { fetchMinionInternalBalances } from '../utils/theGraph';
 import { fetchNativeBalance } from '../utils/tokenExplorerApi';
 import { formatNativeData } from '../utils/vaults';
-import { fetchSafeDetails } from '../utils/requests';
 import { MINION_TYPES } from '../utils/proposalUtils';
 import { useMetaData } from '../contexts/MetaDataContext';
 import { DAO_BOOKS_HOST } from '../data/boosts';
@@ -87,38 +84,22 @@ const MinionVault = ({ overview, customTerms, daoVaults }) => {
       const targetChainId = vaultMatch.foreignChainId || daochain;
       if (vaultMatch.safeAddress) {
         try {
-          const safe = await fetchSafeDetails(
-            chainByID(daochain).networkAlt || chainByID(daochain).network,
-            vaultMatch.safeAddress,
-          );
-          if (safe) {
-            vaultMatch.isMinionModule = safe.modules.includes(
-              Web3Utils.toChecksumAddress(vaultMatch.address),
-            );
-            setSafeDetails(safe);
-          }
+          const safe = await fetchSafeDetails({
+            chainID: daochain,
+            safeAddress: vaultMatch.safeAddress,
+            minionAddress: vaultMatch.address,
+          });
+          setSafeDetails(safe);
           if (vaultMatch.foreignSafeAddress) {
-            const foreignSafe = await fetchSafeDetails(
-              chainByID(vaultMatch.foreignChainId).networkAlt ||
-                chainByID(vaultMatch.foreignChainId).network,
-              vaultMatch.foreignSafeAddress,
-            );
-            if (foreignSafe) {
-              await foreignSafe.modules.every(async m => {
-                const validModule = await isAmbModule(
-                  m,
-                  vaultMatch.safeAddress,
-                  daochain,
-                  vaultMatch.foreignChainId,
-                );
-                if (validModule) {
-                  foreignSafe.ambModuleAddress = m;
-                  return false;
-                }
-                return true;
-              });
-              setForeignSafeDetails(foreignSafe);
-            }
+            const foreignSafe = await fetchSafeDetails({
+              chainID: vaultMatch.foreignChainId,
+              safeAddress: vaultMatch.foreignSafeAddress,
+              ambController: {
+                chainId: daochain,
+                address: vaultMatch.safeAddress,
+              },
+            });
+            setForeignSafeDetails(foreignSafe);
           }
         } catch (error) {
           console.error(error);
